+++ src/collective/fingerpointing/logger.py
@@ -16,7 +16,7 @@ eventlog = getattr(getConfiguration(), 'eventlog', None)
 
 if eventlog is None:
     # we are running tests
-    logfile = AUDITLOG
+    logfile = os.path.join('.', AUDITLOG)
 else:
     try:
         logpath = eventlog.handler_factories[0].instance.baseFilename
@@ -24,13 +24,17 @@ else:
         logfile = os.path.join(logfolder, AUDITLOG)
     except AttributeError:
         # we are in the debug console
-        logfile = AUDITLOG
+        logfile = os.path.join('.', AUDITLOG)
 
 logger = logging.getLogger(PROJECTNAME)
 logger.setLevel(logging.INFO)
 logger.info('Start logging audit information to ' + AUDITLOG)
 
-handler = logging.FileHandler(logfile)
+# support automatic rotation of audit log files at timed intervals
+# we can later implement a way to make this configurable
+handler = logging.handlers.TimedRotatingFileHandler(
+    logfile, when='midnight', backupCount=30)
+
 formatter = logging.Formatter(FORMAT)
 handler.setFormatter(formatter)
 logger.addHandler(handler)