+++ src/collective/fingerpointing/tests/test_view.py
@@ -26,6 +26,13 @@ class LogViewTestCase(unittest.TestCase):
         with self.assertRaises(Unauthorized):
             self.portal.restrictedTraverse('@@fingerpointing-audit-log')
 
+    def test_get_audit_log_files(self):
+        audit_log_files = self.view._get_audit_log_files
+        # we have at least one log file to process
+        self.assertGreaterEqual(len(audit_log_files), 1)
+        # the first file is the current audit log
+        self.assertEqual('./audit.log', audit_log_files[0])
+
     def test_log_tail(self):
         from Products.PlonePAS.events import UserLoggedOutEvent
         from zope.event import notify